git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@2675 01821946-851b-0410-b7f4-f1477c70e57a

+++ ogl/ogl_texture.cpp
@@ -845,7 +845,7 @@ return 0;
 
 void CBitmap::UnlinkTexture (void)
 {
-if (Texture () && (Texture ()->Handle () == (GLuint) -1)) {
+if (Texture () && (Texture ()->Handle () == -1)) {
 	Texture ()->SetHandle (0);
 	SetTexture (NULL);
 	}
